# optimize symbolic-related updates in graphs