llama 8b 6gpus:

```
enqueue in   2.01 ms
total   6.99 ms, 143.01 tok/s, 3084.66 GB/s, param 2299.00 GB/s
```
vs
```
enqueue in   4.65 ms
total   7.85 ms, 127.42 tok/s, 2748.32 GB/s, param 2048.32 GB/s
```

llama 8b 4gpus:
```
enqueue in   1.39 ms
total   7.28 ms, 137.34 tok/s, 2696.59 GB/s, param 2207.06 GB/s
```
vs
```
enqueue in   3.14 ms
total   7.30 ms, 136.99 tok/s, 2689.75 GB/s, param 2201.46 GB/s
```